Upgrade Buttons
Fixed how buttons work. Added a client property "price" to each button, way clearner way of handling things. When the cookie is clicked, it checks every button to see if it can be purchased and makes it change its sprite accordingly, and when a item is purchased it checks every item if it can no longer be purchased and updates accordingly. Very proud of how this works now, and will be a lot easier to add new upgrades and such. Way more readable too, im just hype all around.